sync

+++ models/consignment_note_in.go
@@ -5,9 +5,9 @@ import "time"
 //go:generate go run github.com/objectbox/objectbox-go/cmd/objectbox-gogen
 
 type ConsignmentNoteIn struct {
-	Id                          int64        `json:"id"`
+	Id                          uint64        `json:"id"`
 	ExtId                       string       `json:"ext_id"`
-	AppId                       string        `json:"app_id"`
+	AppId                       string       `json:"app_id"`
 	Date                        time.Time    `json:"date"`
 	Number                      string       `json:"number"`
 	HarvestType                 *HarvestType `json:"harvetst_type" objectbox:"link"`
@@ -18,11 +18,11 @@ type ConsignmentNoteIn struct {
 	Sender                      *Storage     `json:"sender" objectbox:"link"`
 	MateriallyResponsiblePerson *AppUser     `json:"materially_responsible_person" objectbox:"link"`
 	ResponsiblePerson           *AppUser     `json:"responsible_person" objectbox:"link"`
-	Gross                       float32      `json:"gross"`
-	Tare                        float32      `json:"tare"`
-	Net                         float32      `json:"net"`
-	Humidity                    float32      `json:"humidity"`
-	Weediness                   float32      `json:"weediness"`
+	Gross                       float64      `json:"gross"`
+	Tare                        float64      `json:"tare"`
+	Net                         float64      `json:"net"`
+	Humidity                    float64      `json:"humidity"`
+	Weediness                   float64      `json:"weediness"`
 	IsDeleted                   bool         `json:"is_deleted"`
 	CreatedAt                   time.Time    `json:"created_at"`
 	UpdatedAt                   time.Time    `json:"updated_at"`
